Some modifications that has been requested on AdminDashboard like making the reutn as double instead of int and retrive the readonly data using AsNoTracking

+++ Features/Analytics/GetAdminDashboard/Queries/GetTotalAttemptsQuery.cs
@@ -2,6 +2,6 @@
 
 namespace exam_system.Features.Analytics.GetAdminDashboard.Queries
 {
-    public record GetTotalAttemptsQuery : IRequest<int>;
+    public record GetTotalAttemptsQuery : IRequest<double>;
     
 }
